@@ -4,7 +4,7 @@ version: 2.0.1
 homepage: https://github.com/fayeed/flutter_mentions
 
 environment:
-  flutter: "3.19.6"
+  flutter: "3.24.0"
   sdk: '>=2.19.0 <4.0.0'
 
 dependencies:
