@@ -3,7 +3,7 @@ module.exports = {
     rules: {
         'header-max-length': [2, 'always', 100],
 
-        'type-enum': [2, 'always', ['build', 'chore', 'ci', 'docs', 'feat', 'feature', 'fix', 'refactor', 'style', 'test']],
+        'type-enum': [2, 'always', ['build', 'chore', 'ci', 'docs', 'feat', 'feature', 'fix', 'refactor', 'style', 'test', 'feat!', 'fix!']],
         'type-empty': [2, 'never'],
         'type-case': [2, 'always', 'lower-case'],
 
