@@ -1,11 +1,9 @@
 import 'package:app/core/domain/chat/chat_repository.dart';
 import 'package:app/core/domain/chat/entities/guild.dart';
 import 'package:app/core/domain/chat/entities/guild_room.dart';
-import 'package:app/core/service/wallet/wallet_connect_service.dart';
 import 'package:app/injection/register_module.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:web3modal_flutter/web3modal_flutter.dart';
 
 part 'check_guild_room_roles_bloc.freezed.dart';
 
@@ -24,8 +22,10 @@ class CheckGuildRoomRolesBloc
     CheckGuildRoomRolesEventFetch event,
     Emitter emit,
   ) async {
-    final userWalletAddress =
-        getIt<WalletConnectService>().w3mService.session?.address ?? '';
+    // final userWalletAddress =
+    //     getIt<WalletConnectService>().w3mService.session?.address ?? '';
+    // TODO: FIX WALLET MIGRATION
+    const userWalletAddress = '';
     emit(CheckGuildRoomRolesState.loading());
     final guild = await _getGuildDetail();
     List<GuildRolePermission> guildRolePermissions = [];
