@@ -3,8 +3,8 @@ import 'package:app/core/service/wallet/wallet_connect_service.dart';
 import 'package:app/injection/register_module.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:web3modal_flutter/services/w3m_service/i_w3m_service.dart';
-import 'package:web3modal_flutter/web3modal_flutter.dart';
+import 'package:reown_appkit/modal/i_appkit_modal_impl.dart';
+import 'package:reown_appkit/reown_appkit.dart';
 
 part 'wallet_bloc.freezed.dart';
 
@@ -56,7 +56,7 @@ class WalletBloc extends Bloc<WalletEvent, WalletState> {
       );
     }
     // Case 0: init error
-    if (w3mService.status == W3MServiceStatus.error) {
+    if (w3mService.status == ReownAppKitModalStatus.error) {
       return emit(
         state.copyWith(
           activeSession: w3mService.session,
@@ -116,7 +116,7 @@ class WalletEvent with _$WalletEvent {
 @freezed
 class WalletState with _$WalletState {
   const factory WalletState({
-    W3MSession? activeSession,
+    ReownAppKitModalSession? activeSession,
     required ConnectButtonState state,
   }) = _WalletState;
 }
