@@ -1,6 +1,6 @@
 import 'package:app/core/application/event/get_event_detail_bloc/get_event_detail_bloc.dart';
 import 'package:app/core/application/wallet/wallet_bloc/wallet_bloc.dart';
-import 'package:app/core/domain/event/entities/event.dart';
+import 'package:app/core/domain/event/entities/event.dart' as event_entity;
 import 'package:app/core/domain/event/event_repository.dart';
 import 'package:app/core/domain/payment/entities/payment_account/payment_account.dart';
 import 'package:app/core/domain/payment/input/create_payment_account_input/create_payment_account_input.dart';
@@ -22,7 +22,6 @@ import 'package:app/theme/spacing.dart';
 import 'package:app/theme/typo.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:web3modal_flutter/web3modal_flutter.dart' as web3modal;
 
 class EventSetupDirectCryptoPaymentAccountPage extends StatefulWidget {
   final Chain chain;
@@ -46,8 +45,10 @@ class _EventSetupDirectCryptoPaymentAccountPageState
   @override
   void initState() {
     super.initState();
-    textController.text =
-        context.read<WalletBloc>().state.activeSession?.address ?? '';
+    // textController.text =
+    //     context.read<WalletBloc>().state.activeSession?.address ?? '';
+    // TODO: FIX WALLET MIGRATION
+    textController.text = '';
     textController.addListener(() {
       setState(() {
         isValid = _validatePayeeAddress(textController.text);
@@ -65,7 +66,8 @@ class _EventSetupDirectCryptoPaymentAccountPageState
 
   bool _validatePayeeAddress(String value) {
     try {
-      web3modal.EthereumAddress.fromHex(value);
+      // web3modal.EthereumAddress.fromHex(value);
+      // TODO: FIX WALLET MIGRATION
       return true;
     } catch (e) {
       return false;
@@ -73,15 +75,17 @@ class _EventSetupDirectCryptoPaymentAccountPageState
   }
 
   Future<void> _onSubmit(
-    Event event,
+    event_entity.Event event,
   ) async {
     setState(() {
       isLoading = true;
     });
     _createdPaymentAccount ??= await _createEthereumRelayPaymentAccount(
       selectedChain: widget.chain,
-      userWalletAddress:
-          context.read<WalletBloc>().state.activeSession?.address ?? '',
+      // userWalletAddress:
+      //     context.read<WalletBloc>().state.activeSession?.address ?? '',
+      // TODO: FIX WALLET MIGRATION
+      userWalletAddress: '',
       payeeAddress: _getPayeeAddress(),
     );
     if (_createdPaymentAccount == null) {
@@ -154,8 +158,8 @@ class _EventSetupDirectCryptoPaymentAccountPageState
     );
   }
 
-  Future<Event?> _updateEventWithNewPaymentAccount({
-    required Event event,
+  Future<event_entity.Event?> _updateEventWithNewPaymentAccount({
+    required event_entity.Event event,
     required PaymentAccount newPaymentAccount,
   }) async {
     final result = await getIt<EventRepository>().updateEvent(
@@ -177,17 +181,20 @@ class _EventSetupDirectCryptoPaymentAccountPageState
   Widget build(BuildContext context) {
     return BlocConsumer<WalletBloc, WalletState>(
       listener: (context, state) {
-        if (state.activeSession?.address != null) {
-          textController.text = state.activeSession?.address ?? '';
-          setState(() {
-            isValid = _validatePayeeAddress(textController.text);
-          });
-        }
+        // if (state.activeSession?.address != null) {
+        //   textController.text = state.activeSession?.address ?? '';
+        //   setState(() {
+        //     isValid = _validatePayeeAddress(textController.text);
+        //   });
+        // }
+        // TODO: FIX WALLET MIGRATION
       },
       builder: (context, state) {
         final t = Translations.of(context);
         final colorScheme = Theme.of(context).colorScheme;
-        final userWalletAddress = state.activeSession?.address;
+        // final userWalletAddress = state.activeSession?.address;
+        // TODO: FIX WALLET MIGRATION
+        const userWalletAddress = '';
         final event = context.read<GetEventDetailBloc>().state.maybeWhen(
               orElse: () => null,
               fetched: (event) => event,
@@ -241,8 +248,12 @@ class _EventSetupDirectCryptoPaymentAccountPageState
                         autofocus: true,
                       ),
                       SizedBox(height: Spacing.medium),
+                      // TODO: FIX WALLET MIGRATION
+                      // ignore: unnecessary_null_comparison
                       if (userWalletAddress == null)
                         const ConnectWalletButton(),
+                      // TODO: FIX WALLET MIGRATION
+                      // ignore: unnecessary_null_comparison
                       if (userWalletAddress != null)
                         Opacity(
                           opacity: (isValid && !isLoading) ? 1 : 0.5,
