@@ -8,12 +8,10 @@ import 'package:app/core/presentation/widgets/common/button/linear_gradient_butt
 import 'package:app/core/presentation/widgets/common/dotted_line/dotted_line.dart';
 import 'package:app/core/presentation/widgets/lemon_network_image/lemon_network_image.dart';
 import 'package:app/core/presentation/widgets/loading_widget.dart';
-import 'package:app/core/service/wallet/wallet_connect_service.dart';
 import 'package:app/core/service/web3/lemonade_relay/lemonade_relay_utils.dart';
 import 'package:app/core/utils/web3_utils.dart';
 import 'package:app/gen/assets.gen.dart';
 import 'package:app/i18n/i18n.g.dart';
-import 'package:app/injection/register_module.dart';
 import 'package:app/theme/color.dart';
 import 'package:app/theme/sizing.dart';
 import 'package:app/theme/spacing.dart';
@@ -22,7 +20,6 @@ import 'package:flutter/material.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
-import 'package:web3modal_flutter/web3modal_flutter.dart' as web3modal;
 
 class ClaimbleRelayPayment {
   final BigInt amount;
@@ -77,7 +74,9 @@ class _ClaimPaymentByNetworkGroupState
                 targetPaymentAccount?.accountInfo?.paymentSplitterContract ??
                     '',
             connectedWalletAddress:
-                getIt<WalletConnectService>().w3mService.session?.address ?? '',
+                // getIt<WalletConnectService>().w3mService.session?.address ?? '',
+                // TODO: FIX WALLET MIGRATION
+                '',
           ),
         );
   }
