@@ -1,18 +1,15 @@
 import 'package:app/core/application/event_tickets/calculate_event_tickets_pricing_bloc/calculate_event_tickets_pricing_bloc.dart';
 import 'package:app/core/domain/payment/entities/payment_account/payment_account.dart';
 import 'package:app/core/presentation/widgets/animation/circular_animation_widget.dart';
-import 'package:app/core/service/wallet/wallet_connect_service.dart';
 import 'package:app/core/utils/payment_utils.dart';
 import 'package:app/core/utils/web3_utils.dart';
 import 'package:app/gen/assets.gen.dart';
 import 'package:app/gen/fonts.gen.dart';
 import 'package:app/i18n/i18n.g.dart';
-import 'package:app/injection/register_module.dart';
 import 'package:app/theme/spacing.dart';
 import 'package:app/theme/typo.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:web3modal_flutter/web3modal_flutter.dart';
 
 class WalletSignaturePendingView extends StatelessWidget {
   final PaymentAccount? selectedPaymentAccount;
@@ -27,8 +24,10 @@ class WalletSignaturePendingView extends StatelessWidget {
   Widget build(BuildContext context) {
     final colorScheme = Theme.of(context).colorScheme;
     final t = Translations.of(context);
-    final walletAddress =
-        getIt<WalletConnectService>().w3mService.session?.address ?? '';
+    // TODO: FIX WALLET MIGRATION
+    const walletAddress = '';
+    // final walletAddress =
+    //     getIt<WalletConnectService>().w3mService.session?.address ?? '';
     return BlocBuilder<CalculateEventTicketPricingBloc,
         CalculateEventTicketPricingState>(
       builder: (context, state) {
