@@ -9,7 +9,7 @@ import 'package:app/theme/spacing.dart';
 import 'package:flutter/material.dart';
 import 'package:graphql_flutter/graphql_flutter.dart';
 import 'package:pull_to_refresh_flutter3/pull_to_refresh_flutter3.dart';
-import 'package:web3modal_flutter/utils/debouncer.dart';
+import 'package:app/core/utils/debouncer.dart';
 
 class UserFarcasterCastsNewsfeed extends StatefulWidget {
   final AirstackFarcasterUser user;
