@@ -6,10 +6,8 @@ import 'package:app/core/presentation/pages/setting/widgets/setting_tile_widget.
 import 'package:app/core/presentation/widgets/common/appbar/lemon_appbar_widget.dart';
 import 'package:app/core/presentation/widgets/theme_svg_icon_widget.dart';
 import 'package:app/core/presentation/widgets/web3/connect_wallet_button.dart';
-import 'package:app/core/service/wallet/wallet_connect_service.dart';
 import 'package:app/core/utils/web3_utils.dart';
 import 'package:app/gen/assets.gen.dart';
-import 'package:app/injection/register_module.dart';
 import 'package:app/router/app_router.gr.dart';
 import 'package:app/theme/sizing.dart';
 import 'package:app/theme/spacing.dart';
@@ -22,7 +20,6 @@ import 'package:app/gen/fonts.gen.dart';
 import 'package:app/i18n/i18n.g.dart';
 import 'package:app/theme/color.dart';
 import 'package:app/core/presentation/widgets/common/button/linear_gradient_button_widget.dart';
-import 'package:web3modal_flutter/web3modal_flutter.dart';
 
 @RoutePage()
 class OnboardingSocialOnChainPage extends StatelessWidget {
@@ -89,16 +86,20 @@ class OnboardingSocialOnChainPage extends StatelessWidget {
                   SizedBox(height: Spacing.medium),
                   ConnectWalletButton(
                     builder: (onPressConnect, connectButtonState) {
-                      final w3mService =
-                          getIt<WalletConnectService>().w3mService;
-                      final userWalletAddress =
-                          connectButtonState == ConnectButtonState.connected
-                              ? w3mService.session?.address
-                              : null;
+                      // TODO: FIX WALLET MIGRATION
+                      // final w3mService =
+                      //     getIt<WalletConnectService>().w3mService;
+                      // final userWalletAddress =
+                      //     connectButtonState == ConnectButtonState.connected
+                      //         ? w3mService.session?.address
+                      //         : null;
+                      const userWalletAddress = '';
                       return SettingTileWidget(
                         radius: LemonRadius.medium,
                         leadingRadius: LemonRadius.xSmall,
                         title: t.common.actions.connectWallet,
+                        // TODO: FIX WALLET MIGRATION
+                        // ignore: unnecessary_null_comparison
                         subTitle: userWalletAddress != null
                             ? Web3Utils.formatIdentifier(userWalletAddress)
                             : t.common.status.notConnected,
