@@ -11,7 +11,7 @@ import 'package:app/theme/typo.dart';
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:slang/builder/utils/string_extensions.dart';
+import 'package:app/core/utils/string_utils.dart';
 
 @RoutePage()
 class NotificationSettingPage extends StatelessWidget {
