@@ -1,6 +1,5 @@
 import 'package:app/core/application/auth/auth_bloc.dart';
 import 'package:app/core/application/wallet/sign_wallet_bloc/sign_wallet_bloc.dart';
-import 'package:app/core/application/wallet/wallet_bloc/wallet_bloc.dart';
 import 'package:app/core/presentation/pages/token_gating/widget/token_gating_wallet_address_box.dart';
 import 'package:app/core/presentation/widgets/bottomsheet_grabber/bottomsheet_grabber.dart';
 import 'package:app/core/presentation/widgets/common/appbar/lemon_appbar_widget.dart';
@@ -12,7 +11,6 @@ import 'package:app/theme/spacing.dart';
 import 'package:app/theme/typo.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:web3modal_flutter/web3modal_flutter.dart';
 
 class TokenGatingWalletVerificationPage extends StatelessWidget {
   const TokenGatingWalletVerificationPage({
@@ -58,9 +56,14 @@ class _View extends StatelessWidget {
             )
             .toList() ??
         [];
-    final clientConnectedAddress =
-        context.watch<WalletBloc>().state.activeSession?.address;
-    final isConnected = clientConnectedAddress != null;
+    // final clientConnectedAddress =
+    //     context.watch<WalletBloc>().state.activeSession?.address;
+
+    // TODO: FIX WALLET MIGRATION
+    const clientConnectedAddress = '';
+    // TODO: FIX WALLET MIGRATION
+    // ignore: unnecessary_null_comparison
+    const isConnected = clientConnectedAddress != null;
     final isVerified = (isConnected &&
         userVerifiedWalletAddresses.contains(clientConnectedAddress));
 
