@@ -1,4 +1,3 @@
-import 'package:app/core/application/wallet/wallet_bloc/wallet_bloc.dart';
 import 'package:app/core/domain/event/entities/event.dart';
 import 'package:app/core/domain/event/repository/event_ticket_repository.dart';
 import 'package:app/core/domain/reward/entities/reward_signature_response.dart';
@@ -22,8 +21,6 @@ import 'package:app/theme/spacing.dart';
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 import 'package:collection/collection.dart';
-import 'package:flutter_bloc/flutter_bloc.dart';
-import 'package:web3modal_flutter/web3modal_flutter.dart' as web3modal;
 
 @RoutePage()
 class EventDetailClaimTokenRewardPage extends StatefulWidget {
@@ -87,12 +84,14 @@ class _EventDetailClaimTokenRewardPageState
       _isClaiming = true;
     });
     try {
-      final walletAddress =
-          context.read<WalletBloc>().state.activeSession?.address;
+      // final walletAddress =
+      //     context.read<WalletBloc>().state.activeSession?.address;
+      // TODO: FIX WALLET MIGRATION
+      const walletAddress = '';
       final txHash = await TokenRewardUtils.claimReward(
         vault: vault!,
         signature: rewardSignatureResponse!.signature!,
-        from: walletAddress ?? '',
+        from: walletAddress,
       );
       final chain = (await getIt<Web3Repository>()
               .getChainById(chainId: vault.network ?? ''))
