@@ -15,10 +15,10 @@ class ConnectivityService {
   Stream<bool> get myStream => _controller.stream;
 
   void initialise() async {
-    ConnectivityResult result = await _connectivity.checkConnectivity();
-    _checkStatus(result);
+    final result = await _connectivity.checkConnectivity();
+    _checkStatus(result.first);
     _connectivity.onConnectivityChanged.listen((result) {
-      _checkStatus(result);
+      _checkStatus(result.first);
     });
   }
 
